--- src/types/Connection.ts
@@ -1,38 +1,0 @@
-export interface DatabaseServeur {
-    id: string;
-    name: string;
-    host: string;
-    port: number;
-    username: string;
-    password?: string;
-    database?: string;
-    type: 'mysql' | 'mariadb';
-    isConnected?: boolean;
-    lastConnected?: Date;
-    defaultDaoPath?: string;
-}
-
-export interface ServeurFormData {
-    name: string;
-    host: string;
-    port: number;
-    username: string;
-    password?: string;
-    database: string;
-    type: 'mysql' | 'mariadb';
-    defaultDaoPath?: string;
-}
-
-export interface TableInfo {
-    name: string;
-    columns: ColumnInfo[];
-}
-
-export interface ColumnInfo {
-    name: string;
-    type: string;
-    nullable: boolean;
-    key: 'PRI' | 'UNI' | 'MUL' | '';
-    default: string | null;
-    extra: string;
-}
